fixed overflows/underflows --> incorrect interest logic

+++ internal/core/policies.go
@@ -15,8 +15,8 @@ const TicksPerYear = 12
 type Policies struct {
 	mutex sync.RWMutex
 	// both values below are integers between 0 and 100
-	interestRate uint8
-	reserveReq uint8
+	interestRate uint32
+	reserveReq uint32
 }
 
 func (p *Policies) Populate() {
@@ -24,13 +24,13 @@ func (p *Policies) Populate() {
 	p.reserveReq = 0
 }
 
-func (p *Policies) GetInterestRate() uint8 {
+func (p *Policies) GetInterestRate() uint32 {
 	p.mutex.RLock()
 	defer p.mutex.RUnlock()
 	return p.interestRate
 }
 
-func (p *Policies) GetReserveReq() uint8 {
+func (p *Policies) GetReserveReq() uint32 {
 	p.mutex.RLock()
 	defer p.mutex.RUnlock()
 	return p.reserveReq
